+++ lib/event_bloc_widgets.dart
@@ -1,0 +1,10 @@
+/// This contains the widgets that go along with the base dart EventBloc
+///
+/// See [RepositoryProvider] and [BlocProvider] for good places to start.
+library event_bloc_widgets;
+
+// export 'package:event_bloc/src/widgets/bloc_provider.dart';
+// export 'package:event_bloc/src/widgets/event_channel.dart';
+// export 'package:event_bloc/src/widgets/repository_provider.dart';
+
+export 'package:event_bloc/event_bloc.dart';